Added missing BayerConvert PV's to fix color prosilica and Manta cameras.
Also modified code to throw errors on invalid BayerConvert values instead of
	quietly ignoring them.
Modified trace so only TraceFlow shows msgs on PSReadStatistics and PSResetTimer
	and TraceIODriver only shows msgs while acquiring images.
Set default for TRSCAN macro to Passive, as the MantaG282B stops providing
images if we periodically process PSResetTimer at the TRSCAN scan rate.
Will need a better solution if we decide we need better timestamps from
the MantaG282B, but will likely not need it as we'll use our EVR for timestamps.